requirements.txt
Django==1.6.2
Fabric==1.10.0
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==1.6
South==0.8.4
Sphinx==1.2.3
coverage==3.7.1
django-ajax-selects==1.3.4
django-extensions==1.3.3
django-filter==0.7
django-nose==1.2
django-pdb==0.4.0
djangorestframework==2.3.13
docutils==0.12
ecdsa==0.11
ipython==2.0.0
mock==1.0.1
model-mommy==1.2.1
nose==1.3.3
paramiko==1.15.1
pillow==2.5.3
pycrypto==2.6
pyreadline==2.0
six==1.3.0
xlrd==0.9.3
xlwt==0.7.5
